Cleanup

+++ gl_shader/WindowInterface.h
@@ -12,7 +12,7 @@ namespace SoftwareGL {
 		// Startup of the window (with GL version).
 		virtual bool Startup(const std::pair<int, int>& gl_version) = 0;
 		// Compute run (called after the event run).
-		virtual bool RunCompute(const float delta_time) = 0;
+		virtual bool RunCompute(const double delta_time) = 0;
 		// Event run will give the event.
 		virtual bool RunEvent(const SDL_Event& event) = 0;
 		// Cleanup after the window has been closed.
