+++ gl_shader/WindowSoftwareGL.cpp
@@ -10,7 +10,7 @@ bool WindowSoftwareGL::Startup(const std::pair<int, int>& gl_version)
 	return (gl_version.first >= 4) && (gl_version.second >= 3);
 }
 
-bool WindowSoftwareGL::RunCompute(const float delta_time)
+bool WindowSoftwareGL::RunCompute(const double delta_time)
 {
 	return true;
 }
